refactor(api): move common types to separate module
- Extract LlmOptions, ConvertRequest, MarkdownResponse, and ConvertResult to markitdown_api/types.py
- Update imports in other modules to use the new location for these types
- Simplify _build_markitdown function by removing prompt parameter
